# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ecrossflow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet:
    get:
      operationId: getWallet
      tags: [wallet]
      summary: Get wallet balance
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Wallet balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"
  /wallet/deposit:
    post:
      operationId: createDeposit
      tags: [wallet]
      summary: Create a deposit request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositRequest"
      responses:
        "201":
          description: Deposit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /wallet/withdraw:
    post:
      operationId: createWithdrawal
      tags: [wallet]
      summary: Create a withdrawal request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequest"
      responses:
        "201":
          description: Withdrawal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /wallet/convert:
    post:
      operationId: convertCurrency
      tags: [wallet]
      summary: Convert between currencies
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConvertRequest"
      responses:
        "200":
          description: Conversion done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConvertResponse"
  /wallet/rates:
    get:
      operationId: getExchangeRates
      tags: [wallet]
      summary: Get current exchange rates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Exchange rates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRates"

  # TRANSACTIONS ENDPOINTS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    WalletBalance:
      type: object
      properties:
        balanceUsd:
          type: number
        balancePending:
          type: number
        balanceReserved:
          type: number
        totalBalance:
          type: number
      required:
        - balanceUsd
        - balancePending
        - balanceReserved
        - totalBalance
    DepositRequest:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
        paymentMethod:
          type: string
          enum: [MONCASH, NATCASH, CARD, BANK_TRANSFER, CRYPTO, PAYPAL]
        reference:
          type: string
        notes:
          type: string
      required:
        - amount
        - currency
        - paymentMethod
    WithdrawalRequest:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
        paymentMethod:
          type: string
          enum: [MONCASH, NATCASH, CARD, BANK_TRANSFER, CRYPTO, PAYPAL]
        destination:
          type: string
      required:
        - amount
        - currency
        - paymentMethod
        - destination
    ConvertRequest:
      type: object
      properties:
        amount:
          type: number
        fromCurrency:
          type: string
        toCurrency:
          type: string
      required:
        - amount
        - fromCurrency
        - toCurrency
    ConvertResponse:
      type: object
      properties:
        convertedAmount:
          type: number
        rate:
          type: number
        fee:
          type: number
      required:
        - convertedAmount
        - rate
        - fee
    ExchangeRates:
      type: object
      properties:
        base:
          type: string
        rates:
          type: object
          additionalProperties:
            type: number
        updatedAt:
          type: string
          format: date-time
      required:
        - base
        - rates
        - updatedAt

    # TRANSACTION SCHEMAS
    Transaction:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [DEPOSIT, WITHDRAWAL, BOARD_PAYMENT, BOARD_RECEIPT, REFERRAL_BONUS, CONVERSION, ACTIVATION_FEE, BOARD_PROMOTION, SYSTEM_FEE]
        amount:
          type: number
        currency:
          type: string
        amountUsd:
          type: number
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED, CANCELLED]
        paymentMethod:
          type: string
          nullable: true
        referenceId:
          type: string
          nullable: true
        fromBoard:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - amount
        - currency
        - amountUsd
        - status
        - createdAt
        - updatedAt
...